[ICU-12663] UnicodeString support nearly 2G UChars & avoid integer overflows; string-to-string case mappings & detect integer overflows; make some calls to uprv_memcpy() and similar a little more robust by casting int32_t to size_t before multiplying X-Commit-URL: https://ssl.icu-project.org/trac/changeset/39288 Former-commit-id: 0385ea8309dff2c71ad1cc65dfad7be622e62924

+++ icu4c/source/common/uarrsort.c
@@ -122,7 +122,7 @@ doInsertionSort(char *array, int32_t length, int32_t itemSize,
         if(insertionPoint<j) {
             char *dest=array+insertionPoint*itemSize;
             uprv_memcpy(pv, item, itemSize);  /* v=array[j] */
-            uprv_memmove(dest+itemSize, dest, (j-insertionPoint)*itemSize);
+            uprv_memmove(dest+itemSize, dest, (j-insertionPoint)*(size_t)itemSize);
             uprv_memcpy(dest, pv, itemSize);  /* array[insertionPoint]=v */
         }
     }
@@ -183,7 +183,7 @@ subQuickSort(char *array, int32_t start, int32_t limit, int32_t itemSize,
         right=limit;
 
         /* x=array[middle] */
-        uprv_memcpy(px, array+((start+limit)/2)*itemSize, itemSize);
+        uprv_memcpy(px, array+(size_t)((start+limit)/2)*itemSize, itemSize);
 
         do {
             while(/* array[left]<x */
@@ -202,9 +202,9 @@ subQuickSort(char *array, int32_t start, int32_t limit, int32_t itemSize,
                 --right;
 
                 if(left<right) {
-                    uprv_memcpy(pw, array+left*itemSize, itemSize);
-                    uprv_memcpy(array+left*itemSize, array+right*itemSize, itemSize);
-                    uprv_memcpy(array+right*itemSize, pw, itemSize);
+                    uprv_memcpy(pw, array+(size_t)left*itemSize, itemSize);
+                    uprv_memcpy(array+(size_t)left*itemSize, array+(size_t)right*itemSize, itemSize);
+                    uprv_memcpy(array+(size_t)right*itemSize, pw, itemSize);
                 }
 
                 ++left;
